Set output path for intermediate artifacts
VS build artifacts, bin/ and obj/ directories are created by default
when building .NET apps. To keep the project structure clean and easy to
access all outputs and build results the intermediate build artifacts
were set to be created in the
../Build/Intermediate/
directory of the DesktopApp project.

NOTE: The same operation should be done to whatever project created in
this solution