Create Login screen
The login screen is the first and default page displayed in the
application after it's start.
It contains two text boxes (one as PasswordBox for User's password) and
the icons presenting the text boxes purpose.
At the bottom it has two buttons each for Login and Register actions.

This commit provides only the View layer.
ViewModel and Model (User, account) are still to be implemented.