Add User's model and his session to models properties
The logged in User needs to have the possibility to view his own data,
which requires extending each model with the user's ID.
This unique ID will be added to the model entity each time a certain
user creates new item.
This unique ID will also be used when pulling the data from the database
so only this certain user's data will be viewed during his session.