Implement full login window functionality
The LoginWindow has a responsibility of loggin and registering the user.
Full implementation of this window uses the PasswordBox extension
solution (see PasswordBoxAssistant.cs file).
ModernButton style is edited to fix the inability to add the Command
binding to a templated control.

NOTE: Previously used LoginView has been deleted, as it will not be used
in the application - LoginWindow now handles it's feature.