Create User session
The User's session is implemented as a Singleton to make it accessible
to each layer operating on the data related to the user. This also
prevents (along with other mechanism intentional or not) the application
from handling more than one User session at a time.

Although Singleton pattern has a disadvantage of introducing a global
variable - increased memory usage - the UserSession is very small so the
consumption will not increase significantly.
The other option, worth to consider at some point, would be to add a
UserSessions table into the database, which would be read for the first
or last entry (depending on the strategy) for the currently logged in
user.