Add BooksListViewTemplate to the Styles resources
The BooksListViewTemplate will be used to dynamically create the list of
Books in the db storage. The template needs to specify the way how each
element will look like and how will be rendered.

NOTE: to simplify the codebase the new template has been placed in the
same file as ModernButton. But as there are now more templates, this
ResourceDictionary file has been renamed.