Books screen - main navigation & widgets (#8)
* Add empty BooksScreen UserControl

* Move BooksScreen to Books subdirectory

The Books screen will contain several UserControls, so to keep it clean
and separated from other screens it has been moved to separate directory:
View/Screens/Books/
where all the UserControls related to the Books screns will be kept.

* Add BooksListViewTemplate to the Styles resources

The BooksListViewTemplate will be used to dynamically create the list of
Books in the db storage. The template needs to specify the way how each
element will look like and how will be rendered.

NOTE: to simplify the codebase the new template has been placed in the
same file as ModernButton. But as there are now more templates, this
ResourceDictionary file has been renamed.

* Add ListView for books

* Implement BooksViewModel

* Add opactiy animation to the book panel widget