Handle book selecting by and display book details
When a book is selected by clicking on it's position on the list, the
BooksScreenViewModel creates new BookDetailsPanel giving it a whole
model of selected book.
It is an object taken directly from the list pulled from database.
The BookDetailsPanel passes this model (BookModel) to the
BookDetailsViewModel object, which uses this model to present it's
properties in the view (using bindings).
So when Books screen is opened, the database get's accessed only once to
pull all the books.