Present the properties of the book in details panel
The presentation contains all the elements that are included in the
database for the book.

The only exception is the RatingColor property which is used in the
model only to contain the logic for the rating color.
This property is marked as NotMapped attribute, as it is not included in
the datbase.